requirements.txt
django==5.2.6
django-health-check==3.20.8
psutil==6.1.1
jira==3.10.5
requests
azure-storage-blob==12.27.1
